Move STX files into lib/smart-transaction
